feat(association, report, user, billing, payment, charge, file importing): add enums; add FileExportDTO and reports; implemented BillingService, ReportingService, AuthService,  PaymentService; enhanced ChargeService

+++ src/main/java/com/runaitec/credimacpato/controller/PaymentController.java
@@ -29,7 +29,7 @@ public class PaymentController {
     }
 
     @GetMapping("/voucher/{voucherId}")
-    public ResponseEntity<List<PaymentResponseDTO>> listByVoucher(@PathVariable Integer voucherId) {
+    public ResponseEntity<List<PaymentResponseDTO>> listByVoucher(@PathVariable Long voucherId) {
         return ResponseEntity.ok(paymentService.listByVoucher(voucherId));
     }
