+++ src/main/java/com/runaitec/credimacpato/dto/payment/PaymentResponseDTO.java
@@ -26,7 +26,7 @@ public class PaymentResponseDTO {
 
     private Long customerId;
 
-    private Integer voucherId;
+    private Long voucherId;
 
     private List<Long> paidItemIds;
 }
