+++ src/main/java/com/runaitec/credimacpato/repository/VoucherItemRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface VoucherItemRepository extends JpaRepository<VoucherItem, Integer> {
+public interface VoucherItemRepository extends JpaRepository<VoucherItem, Long> {
 }
